refactor: extract RecordingSessionStartStatusPresenter.swift from RecordingSessionPresenters.swift
Splits the `RecordingSessionStartStatusPresenter` (~80 lines) out of
RecordingSessionPresenters.swift into its own file. The remaining
Cancel/StopReadiness/StopFailure presenters stay together.

Byte-identical move of lines 4-83. Both files import Combine + Foundation.

RecordingSessionPresenters.swift: 190 → 109 lines. Tests: 667.
